requirements.txt
numpy>=1.24
Pillow>=10.0
scikit-image>=0.21
fastapi>=0.110
uvicorn>=0.27
